+++ src/main/java/students/student_konstantin_karetnikov/Homework/Lesson2/Level7/NoSleepNoFocus.java
